bugfix bluetooth connexion / bluetooth border-color / undefined battery
